Data convert to bytes

+++ Scripts/Modules/Networks/Commands/NetReader.cs
@@ -3,18 +3,18 @@ using TinyMVC.ReactiveFields;
 
 namespace TinyMVC.Modules.Networks.Commands {
     internal sealed class NetReader {
-        public readonly byte group;
-        public readonly byte part;
-        public readonly ushort key;
+        public readonly ushort group;
+        public readonly ushort part;
+        public readonly byte key;
         public readonly List<ActionListener<object>> listeners;
         
-        public NetReader(byte group, byte part, ushort key, ActionListener<object> listener) {
+        public NetReader(ushort group, ushort part, byte key, ActionListener<object> listener) {
             this.group = group;
             this.part = part;
             this.key = key;
             listeners = new List<ActionListener<object>>() { listener };
         }
         
-        public bool IsCurrent(byte groupValue, byte partValue, ushort keyValue) => group == groupValue && part == partValue && key == keyValue;
+        public bool IsCurrent(ushort groupValue, ushort partValue, byte keyValue) => group == groupValue && part == partValue && key == keyValue;
     }
 }
